Clone the credentials when adding Schedling a Remoted data object
We need to attach a cloned version of the authentication data, because:
1. donwloadUrl will delete credentials after final state
2. in a later phase, when attachments are fetched, these need to be reused.
-> If not cloned, the credentials might not be availible for the download of the attachments
Alternative: not delete the credentials after download, but the not always clear when exaclty query may be deleted.
E.g. after import-submission we're quite sure. But what if something goes wrong before that, or a download just takes longer.
The highly aync process makes it complicated
Note: probably some clean up background job might be needed. Needs
perhaps a bit of better thinking