Improvements in error handling, also download fail
When an error occurs, an error instance is inserted in the database, and
this is now linked to the task and job via the task:error predicate.
When a download fails, the error message is now also included when it
can be found or a link to the failed file data object if it exists (wich
will be empty because the download failed, I think it is a bit silly).