Return submission on registration, accept JSON-LD
Registering a submission also returns the submission URI and not only
the job URI. Still need to make this proper JSON-LD.
The status API now also accepts JSON-LD that is going to be enriched
with the same mechanism as the registration. We'll reuse the
authentication later on.
The status API now accepts the submission URI instead of the job URI.