Rewrite function for melding api
This function in the controller was long and difficult to read. Moved a
lot of code into separate reusable functions that throw erros when there
is a problem. Simplified error handling into a single try-catch. Also
used RDFJS triples and a store to carry around triples instead of a non
documented format (the result of jsonld.toRDF).