Reading and writing authentication from org graph
Also the authentication data should be nicely kept in the organisation
graph. Now everything should be strictly written and read from the
organisation graph so that it does not end up in the vendor graph. One
exception is the reading and writing of tasks wich might also change in
the future.